requirements.txt
aiogram>=3.9,<4
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
python-dotenv>=1.0,<2
aiosqlite>=0.20,<0.21
jinja2>=3.1,<3.2
